Fix Detector creating and attaching all gestures twice on first mount (#2914)
## Description

After refactoring, the Detector would create and attach all gestures
twice on the first mount. This made gestures effectively not work on
Android, since two gestures would be attached to the view. JS side would
listen to the events of only the second one, while on the native side,
the second one would always be canceled by the second one.

Note that the gestures don't work when the children of the detector
change, which is probably caused by the Reanimated upgrade with the new
example app.

## Test plan

Check any of the new API examples on Android