[Web] Reset web-exclusive properties when handler is disabled. (#3041)
## Description

As mentioned in [this comment](https://github.com/software-mansion/react-native-gesture-handler/issues/3010#issuecomment-2273906102), it may be confusing that Gesture Handler blocks default events even if it is explicitly disabled. This PR adds styles reset when `enabled` property changes.

Currently it changes 3 things:

1. Changes `touch-action` to default value
2. Changes `user-select` to default value
3. Re-enables context menu usage.

I'm not sure if I missed something. If so, please mention it in a comment.

## Test plan

Tested on newly added example.