fix: touch event absolute position (#3078)
## Description

The PR fixes the issue mentioned
[here](https://github.com/software-mansion/react-native-gesture-handler/issues/3073)
about the incorrect absolute positions on touch events. The view
transformations should not affect the absolute position of the touch, so
the source event should also be passed to gesture handlers and used for
calculations. It applies to `ACTION_DOWN`, `ACTION_MOVE`, and
`ACTION_UP` motion events.

<!--
Description and motivation for this PR.

Include 'Fixes #<number>' if this is fixing some issue.
-->

## Test plan

I've checked on android device for repro provided in the above-mentioned
issue with some scroll offset.

<!--
Describe how did you test this change here.
-->