feat: Clangd support (#3069)
## Description

This pull request introduces `clangd` support. You can read more about
it in
[this](https://github.com/software-mansion/react-native-reanimated/pull/6430)
awesome PR that bring it to reanimated. For now it supports iOS only as
I am not sure if we need it for android here. To make it work, you have
build `FabricExample` or `MacOSExample` first.

<!--
Description and motivation for this PR.

Include 'Fixes #<number>' if this is fixing some issue.
-->


<!--
Describe how did you test this change here.
-->