[Web] Remove `findNodeHandle` function. (#3127)
## Description

`findNodeHandle` is a utility function that allows us to get native view from React component. The problem is, that under the hood it uses `findDOMNode`, which is deprecated and [will be removed in React 19](https://react.dev/reference/react-dom/findDOMNode). To get rid of this function, I've written our implementation that meets our requirements. 

Also, for new API we had to introduce a little trick (thanks @j-piasecki) - we add wrapper `div` with `display: contents;` style and we pass a ref to this element. With that change, it is much easier to get required `HTMLElement` - we simply have to get `div` from `ref` and find first child (preferably the one without `display: contents;`)

>[!WARNING]
> You may find tat in `StrictMode` most of new API examples doesn't work. That's not really the case. If you check example from [this comment](https://github.com/software-mansion/react-native-gesture-handler/pull/3127#issuecomment-2376061542) it works fine. What really don't work in these examples are the animations.

## Test plan

Tested that example app works and no `findNodeHandle` error is thrown in `StrictMode`