Infer accessibility label from children in RNGestureHandlerButton on iOS (#3290)
## Description

When using RectButton and friends without an explicit
`accessibilityLabel` prop set, the accessibility label is empty on iOS,
whereas on Android it's correctly inferred from its children like other
react-native components work. I tracked this down to this component not
inheriting from `RCTView` where this algorithm is defined. So to fix
this, I simply vendored that function with a new prefix.

## Test plan

Validated via Maestro Studio.