Fix `TouchableNativeFeedback` regression (#3294)
## Description

#3260 introduced a regression to the `Gesture Handlers`'s
`TouchableNativeFeedback` component.
The component used to be a direct reference to the
`TouchableNativeFeedback`, which has a [couple of static
methods](https://reactnative.dev/docs/touchablenativefeedback#methods).
When it was replaced with a function component, the static methods got
removed, leaving only the render function of the
`TouchableNativeFeedback`.

## Test plan

See how the type errors in the
`example/src/release_tests/touchables/index.tsx` are resolved.