Update the generated specs for the old architecture to work on 0.77 (#3301)
## Description


https://github.com/software-mansion/react-native-gesture-handler/pull/3278
and
https://github.com/software-mansion/react-native-gesture-handler/pull/3277
added support for RN 0.77 where the result of generated java files has
changed but those weren't updated in the PRs. This should've been caught
by the CI but the `package.json` files weren't listed as paths that
would trigger it. This seems wrong, as I would imagine we would like the
CI to run when changing dependencies, especially React Native.

## Test plan

Build the example app (0.75 old arch) and Fabric example app (0.77, both
old and new arch)