Don’t pass gesture handler wrapper props to child (#3343)
## Description

When using useEvent from Reanimated together with PureNativeButton, I noticed that I would receive duplicate events. Upon further investigation I found that this was because createHandler renames the handlers and when reanimated attaches the worklet it does so twice because the same handler is available under different props.

This PR fixes this problem by simply not passing wrapper specific props to the child.

## Test plan

<!--
Describe how did you test this change here.
-->