Update how handler configuration is handled on Android (#3484)
## Description

All handler factories were stored in the `RNGestureHandlerModule` which
always seemed like an odd place to keep them in to me. This PR moves
each factory to be a nested class of the relevant gesture handler. This
gives us private access to the handler inside the factory, which in turn
allows to get rid of the setter methods and assign directly instead.
Those were with use since RNGH was written in java, but we have Kotlin
now. It's time to let go and use proper syntax.

It also adds default constants for each config property that didn't have
it before.

## Test plan

Example app