Fix buttons getting stuck in scroll views during overscroll (#3509)
## Description

Addresses the issue reported in
https://github.com/software-mansion/react-native-gesture-handler/pull/3350.

Conceptually, it's very similar to the above PR. The main difference is
the timing - the above PR cancelled the button during `onReset`, while
this one adds `onFail` method (analogous to `onCancel`). This method is
called just before the state change, while the timing of `onReset` is
dependent on the incoming touch stream and handling of other gestures.

With this change, I wasn't able to reproduce the issue anymore, though
@latekvo mentioned that he was able to get into a situation where none
of `onCancel`, `onFail`, and `onReset` were called. This would likely
manifest in the same way, but it looks like an entirely different
problem. Here, the gesture was canceled, while the underlying button was
still kept as a touch responder, while in the above situation, the
gesture would have never finished (or never activated).

## Test plan

Tested on the linked reproducer and on the Example app