Make base `GestureHandler` on Android non generic (#3491)
## Description

Depends on
https://github.com/software-mansion/react-native-gesture-handler/pull/3489

On Android, `GestureHandler` was a generic class specialized by the
concrete type of the handler. The only place it was used was `applySelf`
method, which allowed for inherited methods to return the concrete
handler:

```
tapHandler.setEnabled(true) // returns TapGestureHandler instead of GestureHandler
```

...but we don't use this pattern anywhere. It made much more sense in
Java, than it does in Kotlin, where we can simply use `apply`, `let`,
etc. when there's a need to do multiple operations on the same object.
Now, it simply makes the logic dealing with gesture handlers more
complicated as it also needs to keep track of the generic type, which in
most cases is `*` so it doesn't give us more type safety.

This PR makes it so that the base `GestureHandler` class is no longer
generic, which should make life easier when dealing with non-concrete
gesture handlers in the future.

## Test plan

Build example apps