Copy `README.md` to package (#3558)
## Description

After moving into monorepo setup main `README.md` was left in the root of the repository. It looks good on `github`, but unfortunately on `npm` you can no longer see readme since it is not present in the package.

## Test plan

Run `npm pack` and see that `README.md` is present in the package.