Fix combo example crashing. (#3587)
## Description

Combo example is crashing due to an invalid prop being passed down to
one of the components.
This code was always invalid, but wasn't crashing up until recently.
The invalid part of the code was removed in this PR - this removal does
not change any expected behaviour.

## Test plan

Open combo example