Update how Gesture Handler exposes `setGestureState` to the Reanimated UI runtime (#3207)
## Description

Changes how `setGestureState` is exposed to the UI runtime. Instead of
the weird conditionally adding Reanimated as a dependency on Android and
the weird cast on iOS it uses `_WORKLET_RUNTIME` const injected by
Reanimated into the JS runtime. This allows Gesture Handler to decorate
the UI runtime without direct dependencies between the libraries.

The new approach relies on two methods being added to the global object:
- `_setGestureStateAsync` on the JS runtime
- `_setGestureStateSync` on the UI runtime

which allows for state manipulation also from the JS thread.

The basic example has been modified to easily test the new
functionality.

> [!CAUTION]
> This works only on the New Architecture (and breaks the old one)

## Test plan

Test the expo example app and the modified basic example app