Add `override` keyword to `TypeScript` part of codebase (#3667)
## Description

This PR adds `override` keyword to part of our codebase that is written in `TypeScript`. I think it will improve readability and some cases will be easier to understand.

It also enables `noImplicitOverride` flag in ts compiler, which means that `override` keyword is mandatory when overriding a method. 

## Test plan

Run example app