Add missing types (#3731)
## Description

While writing examples I noticed that some types would be useful. This
includes types for each gesture and their events. This PR adds those
types.

**Important**: The v3 gesture types take over the names of old v2 types,
the v2 types are getting a `Legacy` prefix. This change will have to be
added to the migration guide. Added this as a
[task](https://github.com/orgs/software-mansion/projects/31/views/1?pane=issue&itemId=132253302)
to roadmap.

## Test plan

`yarn lint-js`