Fix SSR error with `useLayoutEffect` (#3752)
## Description
Using `useLayoutEffect` in combination with SSR results in a warning.
This PR adds
[isomorphicLayoutEffect](https://github.com/reduxjs/react-redux/blob/7e2fdd4ee2021e4282e12ba9fc722f09124e30cd/src/utils/useIsomorphicLayoutEffect.ts#L36),
which acts as a `useEffect` when running on server and as
`useLayoutEffect` everywhere else. This fixes
https://github.com/software-mansion/react-native-gesture-handler/issues/3375.

Note: This may cause #2920 to happen when running SSR, however as we
couldn't replicate the issue on main, probably due to same changes to
view recycling within react native, I think it would even be ok to
simply revert
https://github.com/software-mansion/react-native-gesture-handler/pull/2925,
and simply only `useEffect`, so it shouldn't be an issue.

I decided to use `isomorphicLayoutEffect` instead of reverting to ensure
correctness.

## Test plan

See that there is no warning mentioned in #3375